+++ spec/files/project_test/sub4/subsub2/file.cpp
@@ -1,0 +1,6 @@
+#include <iostream>
+
+int main(int argc, char const *argv[]) {
+    /* code */
+    return 0;
+}